cleanup(sinsp): introduce sinsp_table_owner
